// src/app/interfaces/reasons.ts
export interface Reasons {
    bookingID: number,
    bookingStatus: string,
    bookingStatusRemarks: string,
    createdBy: string,
    modifiedBy: string,
    approverID: number,
    approverType: string,
    reasonID: number,
    providerName?: any;
    emailTo?: any;
    userName?: any;
    hashMoveBookingNum?: any;
    reasonText?: any;
    phoneTo?: any;
    userCountryPhoneCode?: number
}